Extract and return realised sampling time distribution

Extract and return realised sampling time distribution

getSamplingTimeDist(
  record,
  burnin = 0.5,
  maxi = 2,
  numBins = 20,
  show.plot = F
)

Arguments

record

MCMC output produced by inferTTree

burnin

Proportion of the MCMC output to be discarded as burnin

maxi

Maximum generation time to consider

numBins

Number of time bins to compute and display distribution

show.plot

Show a barplot of the distribution

Value

Vector of times between becoming infected and becoming sampled in the posterior