Create a transmission tree compatible with the provided phylogenetic tree

Create a transmission tree compatible with the provided phylogenetic tree

makeCTreeFromPTree(
  ptree,
  off.r = NA,
  off.p = NA,
  neg = NA,
  pi = NA,
  w.shape = NA,
  w.scale = NA,
  ws.shape = NA,
  ws.scale = NA,
  T = NA,
  optiStart = 0
)

Arguments

ptree

Phylogenetic tree

off.r

First parameter of the negative binomial distribution for offspring number

off.p

Second parameter of the negative binomial distribution for offspring number

neg

the within-host effective population size (Ne) timesgeneration duration (g)

pi

probability of sampling an infected individual

w.shape

Shape parameter of the Gamma probability density function representing the generation time

w.scale

Scale parameter of the Gamma probability density function representing the generation time

ws.shape

Shape parameter of the Gamma probability density function representing the sampling time

ws.scale

Scale parameter of the Gamma probability density function representing the sampling time

T

Date when process stops (this can be Inf for fully simulated outbreaks)

optiStart

Method used to optimised colored tree (0=none, 1=slow, 2=fast)

Value

A minimal non-zero probability phylogenetic+transmission tree, or an optimised version if parameters are provided